Internal: Move base control files to controls folder [ED-17488]
