+++ modules/atomic-widgets/controls/base/atomic-control-base.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Elementor\Modules\AtomicWidgets\Base;
+namespace Elementor\Modules\AtomicWidgets\Controls\Base;
 
 use JsonSerializable;
 
@@ -62,3 +62,4 @@ abstract class Atomic_Control_Base implements JsonSerializable {
 		];
 	}
 }
+
