+++ modules/atomic-widgets/controls/types/svg-control.php
@@ -2,7 +2,7 @@
 
 namespace Elementor\Modules\AtomicWidgets\Controls\Types;
 
-use Elementor\Modules\AtomicWidgets\Base\Atomic_Control_Base;
+use Elementor\Modules\AtomicWidgets\Controls\Base\Atomic_Control_Base;
 use Elementor\Modules\AtomicWidgets\Utils\Image\Image_Sizes;
 
 if ( ! defined( 'ABSPATH' ) ) {
