+++ modules/atomic-widgets/elements/atomic-tabs/atomic-tab.php
@@ -12,7 +12,7 @@ use Elementor\Modules\AtomicWidgets\PropTypes\Classes_Prop_Type;
 use Elementor\Modules\AtomicWidgets\Elements\Atomic_Paragraph\Atomic_Paragraph;
 use Elementor\Modules\AtomicWidgets\PropTypes\Attributes_Prop_Type;
 use Elementor\Modules\AtomicWidgets\PropTypes\Color_Prop_Type;
-use Elementor\Modules\AtomicWidgets\Render_Context;
+use Elementor\Modules\AtomicWidgets\Elements\Render_Context;
 use Elementor\Modules\AtomicWidgets\PropTypes\Background_Prop_Type;
 
 if ( ! defined( 'ABSPATH' ) ) {
