+++ modules/atomic-widgets/elements/has-atomic-base.php
@@ -3,8 +3,8 @@
 namespace Elementor\Modules\AtomicWidgets\Elements;
 
 use Elementor\Element_Base;
-use Elementor\Modules\AtomicWidgets\Base\Atomic_Control_Base;
-use Elementor\Modules\AtomicWidgets\Base\Element_Control_Base;
+use Elementor\Modules\AtomicWidgets\Controls\Base\Atomic_Control_Base;
+use Elementor\Modules\AtomicWidgets\Controls\Base\Element_Control_Base;
 use Elementor\Modules\AtomicWidgets\Controls\Section;
 use Elementor\Modules\AtomicWidgets\PropsResolver\Render_Props_Resolver;
 use Elementor\Modules\AtomicWidgets\PropTypes\Contracts\Prop_Type;
