+++ modules/atomic-widgets/elements/loader/frontend-assets-loader.php
@@ -1,5 +1,5 @@
 <?php
-namespace Elementor\Modules\AtomicWidgets\Loader;
+namespace Elementor\Modules\AtomicWidgets\Elements\Loader;
 
 use Elementor\Utils;
 
@@ -41,3 +41,4 @@ class Frontend_Assets_Loader {
 		);
 	}
 }
+
