+++ modules/atomic-widgets/elements/template-renderer/template-renderer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Elementor\Modules\AtomicWidgets\TemplateRenderer;
+namespace Elementor\Modules\AtomicWidgets\Elements\TemplateRenderer;
 
 use Elementor\Utils;
 use ElementorDeps\Twig\Environment;
@@ -60,3 +60,4 @@ class Template_Renderer {
 		return $this->env->render( $name, $context );
 	}
 }
+
