+++ modules/atomic-widgets/styles/cache-validity/cache-validity.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Elementor\Modules\AtomicWidgets\CacheValidity;
+namespace Elementor\Modules\AtomicWidgets\Styles\CacheValidity;
 
-use Elementor\Modules\AtomicWidgets\CacheValidity\Cache_Validity_Item;
+use Elementor\Modules\AtomicWidgets\Styles\CacheValidity\Cache_Validity_Item;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // Exit if accessed directly.
@@ -78,3 +78,4 @@ class Cache_Validity {
 		return $cache_item->get( $keys );
 	}
 }
+
