+++ modules/components/styles/component-styles.php
@@ -4,7 +4,7 @@ namespace Elementor\Modules\Components\Styles;
 
 use Elementor\Core\Base\Document;
 use Elementor\Core\Utils\Collection;
-use Elementor\Modules\AtomicWidgets\CacheValidity\Cache_Validity;
+use Elementor\Modules\AtomicWidgets\Styles\CacheValidity\Cache_Validity;
 use Elementor\Modules\AtomicWidgets\Utils\Utils;
 
 /**
